Add lightweight mysql-client repackager
This downloads the macOS mysql client tarballs, which are
monsters in the 100s of MBs, and strips out binaries that
we don't want to keep. It retains the libraries and other files
that are needed for the binaries to run. Then it repackages
the files and publishes them as artifacts.