Add a workflow to generate multirelease linux binaries for mysql and mysqladmin
Currently the mysql and mysqladmin hermit packages for linux are specific to
ubuntu-20.04. There is a libssl incompatibility between ubuntu 20.04 and 22.04,
so the same dynamic binary cannot be used on both Ubuntu versions. Building
statically linked binaries is apparently A Hard Problem. What I'm doing here
is just bundling both ubuntu versions of the binaries and then we pick the
appropriate one on unpack in the hermit package file.